OPENEUROPA-1216: Make sure that triple store docker image retains data after restart.

+++ run.sh
@@ -7,11 +7,21 @@ sed -i 's/exec virtuoso-t +wait +foreground//g' /virtuoso.sh
 /bin/bash /virtuoso.sh
 
 # Import RDF triples.
-virtuoso-t +configfile /virtuoso.ini +wait
-mv $IMPORT_DIR .
-./vendor/bin/robo purge
-./vendor/bin/robo import
+if [ ! -f ".data_imported" ] ;
+then
+    # Start Virtuoso in background.
+    virtuoso-t +configfile /virtuoso.ini +wait
+    mv $IMPORT_DIR .
 
-# Restart Virtuoso in foreground.
-kill $(pidof virtuoso-t)
+    # Import triples.
+    ./vendor/bin/robo purge
+    ./vendor/bin/robo import
+
+    # Create flag file so we won't re-import if container is restarted.
+    touch .data_imported
+
+    # Restart Virtuoso in foreground.
+    kill $(pidof virtuoso-t)
+fi
+
 exec virtuoso-t +configfile /virtuoso.ini +wait +foreground
